\name{dataDomain}
\docType{data}
\alias{dataDomain}
\title{Municipality map for Canton of Zurich}
\description{
This dataset gives the municipality (gemeinde) map for the Canton of Zurich in Switzerland as of 2019, consisting of 162 municipalities.
}
\usage{dataDomain}
\format{A \code{SpatialPolygons} object containing 162 \code{Polygons}.}
\references{
\href{https://statistik.zh.ch/internet/justiz_inneres/statistik/de/daten/Raeumliche_Daten/Basiskarten.html}{https://statistik.zh.ch/} (Visited: 30/05/2019)
}
\keyword{datasets}
