\name{dataGeo}
\docType{data}
\alias{dataGeo}
\title{Simulated geostatistical data}
\description{
  This dataset gives simulated geostatistical data at 200 locations with a normal-distributed response variable and a covariate.
}
\usage{dataGeo}
\format{A \code{SpatialPointsDataFrame} containing 200 observations with \dQuote{lungfunction} as the response variable and a \dQuote{covariate}.}
\keyword{datasets}
