\name{dataLattice}
\docType{data}
\alias{dataLattice}
\title{Simulated lattice data}
\description{
  This dataset gives simulated lattice data at 162 areas with a Poisson-distributed response variable, a covariate and an offset term. It has the same set of polygons as \code{dataDomain}.
}
\usage{dataLattice}
\format{An \code{sf} data frame containing 162 observations with \dQuote{mortality} as the response variable, a \dQuote{covariate} and a \dQuote{pop} as the population offset term.}
\seealso{\code{\link{dataDomain}}}
\keyword{datasets}
