\name{dataDomain}
\docType{data}
\alias{dataDomain}
\title{Municipality map for Canton of Zurich}
\description{
This dataset gives the municipality (gemeinde) map for the Canton of Zurich in Switzerland as of 2019, consisting of 162 municipalities.
}
\usage{dataDomain}
\format{An \code{sf} data frame containing 162 \code{Polygons}.}
\references{
\href{https://www.zh.ch/de/politik-staat/statistik-daten/datenkatalog.html#/details/151@statistisches-amt-kanton-zuerich}{https://www.zh.ch/de/politik-staat/statistik-daten.html} (Visited: 26/06/2021)
}
\keyword{datasets}
